# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/programs": {
      "get": {
        "operationId": "ProgramsController_list",
        "parameters": [
          {
            "name": "supportType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "financial",
                "education",
                "food",
                "health",
                "maternity",
                "children_family",
                "employment",
                "business",
                "housing",
                "legal",
                "safety",
                "disability",
                "mental_wellbeing",
                "emergency"
              ]
            }
          },
          {
            "name": "region",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "tashkent_city",
                "tashkent_region",
                "andijan",
                "bukhara",
                "fergana",
                "jizzakh",
                "kashkadarya",
                "khorezm",
                "namangan",
                "navoi",
                "samarkand",
                "surkhandarya",
                "syrdarya",
                "karakalpakstan"
              ]
            }
          },
          {
            "name": "open",
            "required": false,
            "in": "query",
            "description": "Only programs currently accepting applications",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "org",
            "required": false,
            "in": "query",
            "description": "Organization slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramSummaryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/programs/{id}": {
      "get": {
        "operationId": "ProgramsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramDetailDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/organizations": {
      "get": {
        "operationId": "ProgramsController_organizations",
        "parameters": [
          {
            "name": "section",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "children_family",
                "health",
                "social_support",
                "disability",
                "education_youth",
                "women_girls",
                "food_humanitarian",
                "emergency",
                "faith_zakat",
                "development"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive name search",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDetailDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "programs"
        ]
      }
    },
    "/api/v1/organizations/{slug}": {
      "get": {
        "operationId": "ProgramsController_organization",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetailDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "programs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hamroh API",
    "description": "Verified support programs, eligibility matching and the scripted assistant.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LocalizedTextDto": {
        "type": "object",
        "properties": {
          "uz": {
            "type": "string"
          },
          "ru": {
            "type": "string"
          },
          "en": {
            "type": "string"
          }
        }
      },
      "OrganizationRefDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "verification": {
            "type": "string",
            "enum": [
              "VERIFIED",
              "UNVERIFIED"
            ]
          }
        },
        "required": [
          "slug",
          "name",
          "verification"
        ]
      },
      "ProgramSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "summary": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "supportTypes": {
            "description": "Support type vocabulary values",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requiredDocuments": {
            "description": "Document vocabulary values",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "applicationStatus": {
            "type": "string",
            "enum": [
              "OPEN",
              "CLOSED",
              "UNKNOWN"
            ]
          },
          "deadline": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "statusVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "isDraft": {
            "type": "boolean"
          },
          "origin": {
            "type": "string",
            "enum": [
              "SAMPLE",
              "RESEARCHED"
            ],
            "description": "SAMPLE = invented prototype data; RESEARCHED = collected from a public source"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Page the program information was collected from"
          },
          "regions": {
            "description": "Region vocabulary values; empty means nationwide",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetCircumstances": {
            "description": "Circumstance vocabulary values the program is designed for",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRefDto"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "supportTypes",
          "genders",
          "requiredDocuments",
          "applicationStatus",
          "deadline",
          "statusVerifiedAt",
          "isDraft",
          "origin",
          "sourceUrl",
          "regions",
          "targetCircumstances",
          "organization"
        ]
      },
      "ProgramDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "summary": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "supportTypes": {
            "description": "Support type vocabulary values",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requiredDocuments": {
            "description": "Document vocabulary values",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "applicationStatus": {
            "type": "string",
            "enum": [
              "OPEN",
              "CLOSED",
              "UNKNOWN"
            ]
          },
          "deadline": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "statusVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "isDraft": {
            "type": "boolean"
          },
          "origin": {
            "type": "string",
            "enum": [
              "SAMPLE",
              "RESEARCHED"
            ],
            "description": "SAMPLE = invented prototype data; RESEARCHED = collected from a public source"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Page the program information was collected from"
          },
          "regions": {
            "description": "Region vocabulary values; empty means nationwide",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetCircumstances": {
            "description": "Circumstance vocabulary values the program is designed for",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRefDto"
          },
          "howToApply": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "ageMin": {
            "type": "number",
            "nullable": true
          },
          "ageMax": {
            "type": "number",
            "nullable": true
          },
          "requiredCircumstances": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "incomeTested": {
            "type": "boolean"
          },
          "source": {
            "type": "string",
            "description": "Where the information came from"
          },
          "researchedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "supportTypes",
          "genders",
          "requiredDocuments",
          "applicationStatus",
          "deadline",
          "statusVerifiedAt",
          "isDraft",
          "origin",
          "sourceUrl",
          "regions",
          "targetCircumstances",
          "organization",
          "ageMin",
          "ageMax",
          "requiredCircumstances",
          "incomeTested",
          "source",
          "researchedAt"
        ]
      },
      "OrganizationDetailDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "verification": {
            "type": "string",
            "enum": [
              "VERIFIED",
              "UNVERIFIED"
            ]
          },
          "description": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "children_family",
                "health",
                "social_support",
                "disability",
                "education_youth",
                "women_girls",
                "food_humanitarian",
                "emergency",
                "faith_zakat",
                "development"
              ]
            }
          },
          "kind": {
            "type": "string",
            "enum": [
              "NGO",
              "INTERNATIONAL",
              "GOVERNMENT"
            ]
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true
          },
          "researchedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "When information was collected from sourceUrl"
          },
          "lastVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "programs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramSummaryDto"
            }
          }
        },
        "required": [
          "slug",
          "name",
          "verification",
          "description",
          "categories",
          "sections",
          "kind",
          "website",
          "phone",
          "email",
          "source",
          "sourceUrl",
          "researchedAt",
          "lastVerifiedAt",
          "programs"
        ]
      },
...
    }
  }
}